inter_project/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.5
coverage==6.4.1